==> R/00_all.qmd <==
---
title: "Group 26 - Project" 
author: 
- "Carl Felix Freiesleben - s203521"
- "Emilie Munk - s203538"
- "Josefine Løken - s183784"
- "Judith Tierno Martí - s222869"
- "Sahand Yazdani - s203538"
format:
  html: 
    embed-resources: true
execute: 
  message: FALSE
  warning: FALSE
---

## Load Data
{{< include 01_load.qmd >}}
```{r}
#| include: false
quarto::quarto_render(input = "01_load.qmd",
                      output_format = "html")

file.rename(from = "01_load.html",
            to = "../results/01_load.html")
```
## Clean Data
{{< include 02_clean.qmd >}}
```{r}
#| include: false
quarto::quarto_render(input = "02_clean.qmd", 
                      output_format = "html")

file.rename(from = "02_clean.html",
            to = "../results/02_clean.html")
```
## Augment Data
{{< include 03_aug.qmd >}}
```{r}
#| include: false
quarto::quarto_render(input = "03_aug.qmd", 
                      output_format = "html")

file.rename(from = "03_aug.html",
            to = "../results/03_aug.html")
```
## Description
{{< include 04_describe.qmd >}}
```{r}
#| include: false
quarto::quarto_render(input = "04_describe.qmd", 
                      output_format = "html")

file.rename(from = "04_describe.html",
            to = "../results/04_describe.html")
```
## Analysis 1
{{< include 05_analysis_1.qmd >}}
```{r}
#| include: false
quarto::quarto_render(input = "05_analysis_1.qmd", 
                      output_format = "html")

file.rename(from = "05_analysis_1.html",
            to = "../results/05_analysis_1.html")
```
## Analysis 2
{{< include 06_analysis_2.qmd >}}
```{r}
#| include: false
quarto::quarto_render(input = "06_analysis_2.qmd",
                      output_format = "html")

file.rename(from = "06_analysis_2.html",
            to = "../results/06_analysis_2.html")
```
## PCA
{{< include 07_PCA.qmd >}}
```{r}
#| include: false
quarto::quarto_render(input = "07_PCA.qmd",
                      output_format = "html")

file.rename(from = "07_PCA.html",
            to = "../results/07_PCA.html")
```
## Logistic Modelling
{{< include 08_logistic_modelling.qmd >}}
```{r}
#| include: false
quarto::quarto_render(input = "08_logistic_modelling.qmd",
                      output_format = "html")

file.rename(from = "08_logistic_modelling.html",
            to = "../results/08_logistic_modelling.html")
```
## All
```{r}
#| include: false

file.rename(from = "00_all.html",
            to = "../results/00_all.html")
```
