==> R/01_load.qmd <==
---
title: "Group 26 - Project"
author: 
- "Carl Felix Freiesleben - s203521"
- "Emilie Munk - s203538"
- "Josefine Løken - s183784"
- "Judith Tierno Martí - s222869"
- "Sahand Yazdani - s203538"
format:
  html: 
    embed-resources: true
execute: 
  message: FALSE
  warning: FALSE
---
# Load packages
```{r}
#| label: Load-librariess
#| message: false 
library(tidyverse)
library(base)
library(patchwork)
library(scales)
library(ggridges)
library(patchwork)
library(ggseqlogo)
library(dplyr)
library(table1)
library(quarto)
```


```{r}
#| label: Load-file-and-write-file
#| message: false
# Create the data directory programmatically
#dir.create(path="data")

# _raw data folder, filename, and original data location
raw_dir <- "../data/_raw/"
data_file <- "rhc.csv"
data_loc <- "https://hbiostat.org/data/repo/"


# Checking if file exist in folder
if( !dir.exists(raw_dir) ){
  dir.create(path = raw_dir)
}
if( !file.exists(str_c(raw_dir, data_file)) ){
  download.file(
    url = str_c(data_loc, data_file),
    destfile = str_c(raw_dir, data_file))
}
rhc_data <- read_csv(file = str_c(raw_dir, data_file))

print(str_c(raw_dir,data_file))

# Write the data to data folder
write_csv(x = rhc_data,
          file = "../data/rhc.csv")
```
