==> R/02_clean.qmd <==
---
title: "Group 26 - Project"
author: 
- "Carl Felix Freiesleben - s203521"
- "Emilie Munk - s203538"
- "Josefine Løken - s183784"
- "Judith Tierno Martí - s222869"
- "Sahand Yazdani - s203538"
format:
  html: 
    embed-resources: true
execute: 
  message: FALSE
  warning: FALSE
---

## Clean data

```{r}
#| label: Load-libraries
#| message: false 
library(tidyverse)
library(base)
library(patchwork)
library(scales)
library(ggridges)
library(patchwork)
library(ggseqlogo)
library(dplyr)
library(table1)

rhc_data <- read_csv("../data/rhc.csv")
#| label: Simple-cleaning-chunk-data-wrangling
#copy paste
rhc_clean <- rhc_data |> 
  bind_cols() |> 
  as_tibble() 

# convert death string to binary: 1 = dead 0 = alive
rhc_clean <- rhc_clean |> 
  mutate(death = case_when(
  death == "No" ~ 0,
  death == "Yes" ~ 1))

#Change RHC strings to factor
rhc_clean<- rhc_clean |> 
    mutate(swang1 = case_when(
           swang1 == "No RHC" ~ 0,
           swang1 == "RHC" ~ 1))

#round age to the nearest whole number
rhc_clean <- rhc_clean |> 
  rename(weight = wtkilo1) |> 
  mutate(age = round(age, 0),
         weight = round(weight, 0),
         edu = round(edu, 0),
         temp1 = round(temp1, 1),
         ph1 = round(ph1, 2))

#Remove old ptid and create new ID's
rhc_clean <- rhc_clean |> 
  select(-c(ptid),
         -ends_with("dte"),
         -adld3p) |> 
  rename(ptID= "...1")

```

## Save data

```{r}
#| label: Writing-new-cleaned-csv
#| echo: false
#| eval: true
# Write the data to disk
write_csv(x = rhc_clean,
          file = "../data/rhc_clean.csv")

```
