==> R/03_aug.qmd <==
---
title: "Group 26 - Project"
author: 
- "Carl Felix Freiesleben - s203521"
- "Emilie Munk - s203538"
- "Josefine Løken - s183784"
- "Judith Tierno Martí - s222869"
- "Sahand Yazdani - s203538"
format:
  html: 
    embed-resources: true
execute: 
  message: FALSE
  warning: FALSE
---

# Load packages
```{r}
library(tidyverse)
library(base)
library(patchwork)
library(scales)
library(ggridges)
library(patchwork)
library(ggseqlogo)
library(dplyr)
library(table1)
```

#Augumented data
```{r}
#| label: Augumenting-initiation.
create_binary_column <- function(x) {
  ifelse(x == "Yes", 1, 
  ifelse(x == "No", 0, x))
}

rhc_clean <- read_csv("../data/rhc_clean.csv")
#Creates one column with all diseases that is reason for admission to the hospital, instead of them all being present for all patients.
rhc_clean_long <- rhc_clean |>
    
  mutate(across(resp:ortho, 
                ~create_binary_column(.))) |>
  pivot_longer(cols = resp:ortho,
               names_to = "Diagnosis",
               values_to = "Values")|> 
  filter(Values==1) |> 
  
  group_by(ptID) |> 
  mutate(Diagnosis = paste(unique(Diagnosis), 
                           collapse = ",")) |> 
  ungroup() 
  

#makes one value for patients with multiple diagnosis as reason for admission and delete repeated columns
rhc_aug <- rhc_clean_long |>
  mutate(Diagnosis = case_when(str_detect(Diagnosis, 
                                          pattern = ",",
                                          negate = FALSE) == TRUE ~ "multiple diagnosis", 
                               str_detect(Diagnosis, 
                                          pattern = ",",
                                          negate = FALSE) == FALSE ~ Diagnosis )) |> 
  unique.data.frame() |> 
  mutate(aps1_Interval = cut(aps1, 
                             breaks = c(0, 15, 30, 45, 60, 75, 90,
                                        105, 120, 135, 150), 
                             labels = c("0-15", "15-30", "30-45", 
                                        "45-60", "60-75", "75-90", 
                                        "90-105","105-120", "120-135", 
                                        "135-150"))) |> 
  mutate(age_group = cut(age, 
                         breaks = c(0, 10, 20, 30, 40, 50,
                                    60, 70, 80, 90, Inf), 
                         labels = c("<10", "10-20", "20-30", 
                                    "30-40", "40-50", "50-60", 
                                    "60-70", "70-80", "80-90", 
                                    "> 90"), 
                         include.lowest = TRUE, 
                         na.rm = TRUE)) |> 
  mutate(edu_group = cut(edu,
                         breaks = c(-Inf, 10, 20, Inf), 
                         labels = c("<10 years", 
                                    "between 10-20 years", 
                                    ">20 years"),
                         include.lowest = TRUE,
                         na.rm = TRUE))

#Add a column with the number of comorbidities each patient has.
rhc_aug <- rhc_aug |>
  rowwise() |> 
  mutate(comorbidities = sum(c_across(cardiohx:amihx)))


#simulated seperated and joined datatable

rhc_split1 <- rhc_aug |>
  select(-c(edu_group, comorbidities)) 

rhc_split2 <- rhc_aug |> 
  select(c(ptID, edu_group, comorbidities))

rhc_aug <- inner_join(x = rhc_split1, 
                      y = rhc_split2, 
                      by ="ptID")

```


## Save data

```{r}
#| label: Writing-new-augmented-csv
#| 
# Write the data to disk
write_csv(x = rhc_aug,
          file = "../data/rhc_aug.csv")

```
