==> R/04_describe.qmd <==
---
title: "Group 26 - Project"
author: 
- "Carl Felix Freiesleben - s203521"
- "Emilie Munk - s203538"
- "Josefine Løken - s183784"
- "Judith Tierno Martí - s222869"
- "Sahand Yazdani - s203538"
format:
  html: 
    embed-resources: true
execute: 
  message: FALSE
  warning: FALSE
---
# Load packages
```{r}
library(tidyverse)
library(table1)
```


The dataset we have used comes from Connors et al. (1996): The effectiveness of RHC in the initial care of critically ill patients. It was publicly available on hbiostat.org and downloaded as a cvs file. 
The dataset contains information about patients admitted to the hospital in order to study the effect of right heart catherization (RHC). There are total of  2184 patients in the "disease" group which recieved RHC and 3551 patients in the "control" group without RHC. The control group was artificially created using a propensity matching score. The dataset contains 5735 rows and 63 columns. 

```{r}

rhc_data <- read_csv("../data/rhc.csv")
rhc_aug <- read_csv("../data/rhc_aug.csv")
```

```{r}
rhc_data |> 
  dim()
```

```{r}
rhc_data |> 
  group_by(swang1) |> 
  summarise(n = n())
```

The 63 variables can be divided into categories- Patient information includes general facts about the patients like race, age, gender. Socioeconomic information includes level of income, years of education, and healthcare. Fysiological stats are measurements like temperature, heart rate and blood pressure. Admission diagnosis are conditoins accounting for the reasons of the hospital visit. Comorbidities illness are diseases caused as a direct affect an underlying disease. Hospitalization information like addmission date, discharge date and date of last contact. Lastly, information about death vs. survival and date of last contact. 

A statistical summary of patient information is displayed in table1. Swang1 is the attribute that shows if the patient recieved RHS or not. 
 
```{r}
T1 <- rhc_aug |> 
  mutate(sex = factor(sex),
         swang1 = factor(swang1),
         death = factor(x = death, 
                        levels = c(0,1), 
                        c("Alive","Dead"))) |> 
  table1(x = formula(~ sex + age + race + swang1 | death),
         data = _) 
T1

```


The broad spectrum of attributes allows for exciting analysis of how treatment type, diseases and economic factors influence the survival rate of people.
