==> R/05_analysis_1.qmd <==
---
title: "Group 26 - Project"
author: 
- "Carl Felix Freiesleben - s203521"
- "Emilie Munk - s203538"
- "Josefine Løken - s183784"
- "Judith Tierno Martí - s222869"
- "Sahand Yazdani - s203538"
format:
  html: 
    embed-resources: true
execute: 
  message: FALSE
  warning: FALSE
---
# SOCIOECONOMIC analysis
We conducted a comprehensive socioeconomic analysis focusing on variables such as age, gender, education, race, income, and medical assurance. Plots and summary statistics explore the distribution of deaths and the application of right heart catheterization (RHC) across different subgroups of patients. The analysis provides insights into potential associations between socioeconomic factors and health outcomes.


## Libraries 
```{r}
library(tidyverse)
library(base)
library(patchwork)
library(scales)
library(ggridges)
library(patchwork)
library(ggseqlogo)
library(dplyr)
library(table1)
```

```{r}
rhc_aug <- read_csv("../data/rhc_aug.csv")
```


Color Scheme:
```{r}
"darkolivegreen2"
"darkgoldenrod1"
"red"
```
The variables that we found to be interesting to analyze from a socioeconomic perspective are: 
```{r}
rhc_aug |> 
  select(age, sex, edu, race, income, ninsclas)
```
For all the variables of interest, we focused on plotting the death distribution of patients and if they had been applied an RHC.


## Age

Following plot provides insights into the mortality rates across different ages.
```{r}
# Death in different ages
rhc_aug |>
  ggplot(aes(x = age, 
             fill = factor(death))) +
  geom_bar() +
  labs(title = "Death in different ages",
       x = "Age",
       y = "Number of patients") +
  scale_fill_manual(name = "Death", 
                    values = c("0" = "blue", "1" = "red")) +
  theme_minimal()

```

Following plot provides insights into the mortality rates across different age groups.
```{r}
# Death in different group ages
rhc_aug |>
  ggplot(aes(x = age_group, 
             fill = factor(death))) +
  geom_bar() +
  labs(title = "Death in different ages",
       x = "Age",
       y = "Number of patients") +
  scale_fill_manual(name = "Death", 
                    values = c("0" = "blue", "1" = "red")) +
  theme_minimal()
```

The following chunk is a representation of the percentage of deaths by age group.
```{r}
## Percentage of death by age
rhc_aug |> 
  group_by(age_group, death) |> 
  summarize(count = n()) |> 
  group_by(age_group) |> 
  mutate(percent = count / sum(count) * 100) |> 
  ungroup() |> 
  filter(death == 1) |> 
  select(age_group, percent)
```
Several observations can be made. Firstly, it becomes apparent that mortality rates increase with advancing age. This finding aligns with existing knowledge that older individuals generally have a higher risk of experiencing adverse health outcomes. Secondly, the plot demonstrates that the number of deaths is relatively higher in certain age brackets, indicating the presence of age-specific vulnerabilities. Further research would be needed to fully understand the underlying reasons for these variations.


Following plot provides insights into the RHC appliance rates across different age groups.
```{r}
# Count of RHC by age
rhc_aug |> 
  ggplot(aes(x = age_group, 
             fill = as.factor(swang1))) +
  geom_bar() +
  labs(title = "Distribution of Age by RHC Status",
       x = "Age",
       y = "Count",
       fill = "RHC Status") +
  scale_fill_manual(name= "RHC Status", 
                    values = c("No RHC" = "blue", "RHC" = "red")) +
  theme_minimal()
  
```

The following chunk shows the corresponding percentage of RHC procedures for each age group.
```{r}
## Percentage of RHC by age_group
rhc_aug |> 
  group_by(age_group, swang1) |> 
  summarize(count = n()) |> 
  group_by(age_group) |> 
  mutate(percent = count / sum(count) * 100) |> 
  ungroup() |> 
  filter(swang1 == "RHC") |> 
  select(age_group, percent)
```
This results suggests that middle-aged patients (40-50 and 60-70 age groups) may have higher rates of RHC procedures, potentially due to specific cardiac conditions or clinical indications. Conversely, patients (10-20 age group) and older (> 90 age group) may have lower rates of RHC procedures, indicating a different clinical approach or lower prevalence of conditions requiring RHC. Further research would be needed to understand the underlying reasons for these variations and fully evaluate the appropriateness of RHC procedures.


## Gender

Following plot provides insights into the mortality rates across different genders.
```{r}
# Death females vs. males
rhc_aug |>
  ggplot(aes(x = sex,
             fill = factor(death))) +
  geom_bar() +
  labs(title = "Death females vs. males",
       x = "Gender",
       y = "Number of patients") +
  scale_fill_manual(name = "Death", 
                    values = c("0" = "blue", "1" = "red")) +
  theme_minimal()
```

The following chunk is a representation of the percentage of deaths by gender.
```{r}
# Percentage of death by gender
rhc_aug |> 
  group_by(sex, death) |> 
  summarize(count = n()) |> 
  group_by(sex) |> 
  mutate(percent = count / sum(count) * 100) |> 
  ungroup() |> 
  filter(death == 1) |> 
  select(sex, percent)
```
This results suggest that there may be a slight difference in mortality rates between females and males, with males having a slightly higher percentage of deaths. However, it is important to note that the difference is relatively small and may not be statistically significant. Further research would be needed to fully understand and check the underlying reasons for these variations.



Following plot provides insights into the RHC appliance rates across different genders.
```{r}
# Count of RHC by gender
rhc_aug |> 
  ggplot(aes(x = sex, 
             fill = as.factor(swang1))) +
  geom_bar() +
  labs(title = "Distribution of Gender by RHC Status",
       x = "Gender",
       y = "Number of patients",
       fill = "RHC Status") +
  scale_fill_manual(name= "RHC Status", 
                    values = c("No RHC" = "blue", "RHC" = "red")) +
  theme_minimal()
```

The following chunk shows the corresponding percentage of RHC procedures for each gender.
```{r}
## Percentage of RHC by gender
rhc_aug |> 
  group_by(sex, swang1) |> 
  summarize(count = n()) |> 
  group_by(sex) |> 
  mutate(percent = count / sum(count) * 100) |> 
  ungroup() |> 
  filter(swang1 == "RHC") |> 
  select(sex, percent)
```
This suggest a potential difference in RHC utilization between females and males, with males having a slightly higher percentage of RHC procedures. Further research would be needed to fully understand and check the underlying reasons for these variations. 


## Race

Following plot provides insights into the mortality rates across different races.
```{r}
#Death by race
rhc_aug |> 
  ggplot(aes(x = race, 
             fill = factor(death))) +
  geom_bar() +
  labs(title = "Death in race",
       x = "Race",
       y = "Count",
       fill = "Death") +
  scale_fill_manual(name = "Death", 
                    values = c("0" = "blue", "1" = "red")) +
  theme_minimal()
```

The following chunk is a representation of the percentage of deaths by race.
```{r}
# Percentage of death by race
rhc_aug |> 
  group_by(race, death) |> 
  summarize(count = n()) |> 
  group_by(race) |> 
  mutate(percent = count / sum(count) * 100) |> 
  ungroup() |> 
  filter(death == 1) |> 
  select(race, percent)
```
This suggest a potential variation in mortality rates among different racial groups, with the white population having a slightly higher percentage of deaths. Further research would be needed to fully understand and check the underlying reasons for these variations. 


Following plot provides insights into the RHC appliance rates across different race.
```{r}
rhc_aug |> 
  ggplot(aes(x = race, 
             fill = as.factor(swang1))) +
  geom_bar() +
  labs(title = "Distribution of Race by RHC Status",
       x = "Race",
       y = "Number of patients",
       fill = "RHC Status") +
  scale_fill_manual(name= "RHC Status", 
                    values = c("No RHC" = "blue", "RHC" = "red")) +
  theme_minimal()
```

The following chunk shows the corresponding percentage of RHC procedures by race.
```{r}
## Percentage of RHC by race
rhc_aug |> 
  group_by(race, swang1) |> 
  summarize(count = n()) |> 
  group_by(race) |> 
  mutate(percent = count / sum(count) * 100) |> 
  ungroup() |> 
  filter(swang1 == "RHC") |> 
  select(race, percent)
```
These findings suggest a potential variation in RHC utilization among different racial groups, with the other racial category having the highest percentage of RHC procedures. Further research would be needed to fully understand and check the underlying reasons for these variations.


## Income

Following plot provides insights into the mortality rates across different patients' incomes.
```{r}
#Death by income
rhc_aug |> 
  ggplot(aes(x = income, 
             fill = factor(death))) +
  geom_bar() +
  labs(title = "Death by Income",
       x = "Income",
       y = "Number of patients",
       fill = "Death") +
  scale_fill_manual(name = "Death", values = c("0" = "blue", "1" = "red")) +
  theme_minimal()
```

The following chunk is a representation of the percentage of deaths by income.
```{r}
## Percentage of death by income
rhc_aug |> 
  group_by(income, death) |> 
  summarize(count = n()) |> 
  group_by(income) |> 
  mutate(percent = count / sum(count) * 100) |> 
  ungroup() |> 
  filter(death == 1) |> 
  select(income, percent)
```
This suggest that the lowest income category have the highest percentage of deaths. However, it is important to note that the differences are relatively small and may not be statistically significant. Further research would be needed to fully understand and check the underlying reasons for these variations.


Following plot provides insights into the RHC appliance rates across different incomes.
```{r}
# Count of RHC by income
rhc_aug |> 
  ggplot(aes(x = income, 
             fill = as.factor(swang1))) +
  geom_bar() +
  labs(title = "Distribution of Income by RHC Status",
       x = "Income",
       y = "Number of patients",
       fill = "RHC Status") +
  scale_fill_manual(name = "RHC Status", 
                    values = c("No RHC" = "blue", "RHC" = "red")) +
  theme_minimal()
```

The following chunk shows the corresponding percentage of RHC procedures by income.
```{r}
## Percentage of RHC by income
rhc_aug |> 
  group_by(income, swang1) |> 
  summarize(count = n()) |> 
  group_by(income) |> 
  mutate(percent = count / sum(count) * 100) |> 
  ungroup() |> 
  filter(swang1 == "RHC") |> 
  select(income, percent)
```
This suggests that patients with higher incomes may have better access to healthcare facilities and may be more likely to undergo diagnostic procedures like RHC. Further research would be needed to fully understand and check the underlying reasons for these variations.


## Education

Following plot provides insights into the mortality rates across different education groups.
```{r}
# Death in different group of education
rhc_aug |>
  ggplot(aes(x = edu_group, 
             fill = factor(death))) +
  geom_bar() +
  labs(title = "Death by Education",
       x = "Years of eduaction",
       y = "Number of patients") +
  scale_fill_manual(name = "Death", 
                    values = c("0" = "blue", "1" = "red")) +
  theme_minimal()
```

The following chunk is a representation of the percentage of deaths by education.
```{r}
## Percentage of death by education
rhc_aug |> 
  group_by(edu_group, death) |> 
  summarize(count = n()) |> 
  group_by(edu_group) |> 
  mutate(percent = count / sum(count) * 100) |> 
  ungroup() |> 
  filter(death == 1) |> 
  select(edu_group, percent)
```
Individuals with higher education levels have slightly higher percentages of deaths. It is possible that individuals with higher education levels may have a higher likelihood of seeking medical attention and being diagnosed with underlying health conditions, which could contribute to the observed higher mortality rates. Further research would be needed to fully understand and check the underlying reasons for these variations.


Following plot provides insights into the RHC appliance rates across different education levels.
```{r}
# Count of RHC by Education
rhc_aug |> 
  ggplot(aes((x = edu_group), 
             fill = as.factor(swang1))) +
  geom_bar() +
  labs(title = "Distribution of Education by RHC Status",
       x = "Education",
       y = "Number of patients",
       fill = "RHC Status") +
  scale_fill_manual(name = "RHC Status", 
                    values = c("No RHC" = "blue", "RHC" = "red")) +
  theme_minimal()

```

The following chunk shows the corresponding percentage of RHC procedures by Education.
```{r}
## Percentage of RHC by education
rhc_aug |> 
  group_by(edu_group, swang1) |> 
  summarize(count = n()) |> 
  group_by(edu_group) |> 
  mutate(percent = count / sum(count) * 100) |> 
  ungroup() |> 
  filter(swang1 == "RHC") |> 
  select(edu_group, percent)
```
Individuals with higher education levels have slightly higher percentages of RHC procedures. They may have better access to healthcare, increased health literacy, and a greater understanding of the benefits of diagnostic procedures like RHC. So healthcare providers may be more likely to recommend RHC to patients with higher education levels. Further research would be needed to fully understand and check the underlying reasons for these variations.


## Medical assurance

Following plot provides insights into the mortality rates across different medical assurance.
```{r}
#Death by medical assurance
rhc_aug |> 
  ggplot(aes(x = ninsclas, 
             fill = factor(death))) +
  geom_bar() +
  labs(title = "Death in medical assurance",
       x = "Medical assurance",
       y = "Number of patients",
       fill = "Death") +
  scale_fill_manual(name = "Death", 
                    values = c("0" = "blue", "1" = "red")) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
  
```

The following chunk is a representation of the percentage of deaths by medical assurance.
```{r}
## Percentage of DEATH by medical assurance
rhc_aug |> 
  group_by(ninsclas, death) |> 
  summarize(count = n()) |> 
  group_by(ninsclas) |> 
  mutate(percent = count / sum(count) * 100) |> 
  ungroup() |> 
  filter(death == 1) |> 
  select(ninsclas, percent)
```
Individuals covered by Medicare having the highest percentage of deaths. This can be because Medicare is primarily for individuals aged 65 and older or those with certain disabilities, who may have a higher prevalence of chronic health conditions and therefore a higher risk of mortality. Further research would be needed to fully understand and check the underlying reasons for these variations.


Following plot provides insights into the RHC appliance rates across different medical assurance.
```{r}
# Count of RHC by Medical Assurance
rhc_aug |> 
  ggplot(aes((x = ninsclas), 
             fill = as.factor(swang1))) +
  geom_bar() +
  labs(title = "Distribution of Medical assurance by RHC Status",
       x = "Medical assurance",
       y = "Number of patients",
       fill = "RHC Status") +
  scale_fill_manual(name = "RHC Status", 
                    values = c("No RHC" = "blue", "RHC" = "red")) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

The following chunk shows the corresponding percentage of RHC procedures by medical assurance.
```{r}
## Percentage of RHC by medical assurance
rhc_aug |> 
  group_by(ninsclas, swang1) |> 
  summarize(count = n()) |> 
  group_by(ninsclas) |> 
  mutate(percent = count / sum(count) * 100) |> 
  ungroup() |> 
  filter(swang1 == "RHC") |> 
  select(ninsclas, percent)
```
Individuals without insurance and those with private insurance have higher percentages of RHC procedures. Individuals without insurance may have limited access to healthcare services and may seek RHC procedures when their condition becomes severe. On the other hand, individuals with private insurance may have better access to healthcare facilities and may be more likely to undergo diagnostic procedures like RHC. Further research would be needed to fully understand and check the underlying reasons for these variations.



## Others

```{r}
death_count <- rhc_aug |>
    mutate(death = case_when(
         death == 0 ~ "Alive",
         death == 1 ~ "Dead")) |> 
  # filter(death == 0) |> 
  mutate(income = factor(income, levels = rev(levels(as.factor(income))))) |> 
  ggplot(aes(x = income, 
             fill = as.factor(death))) +
  geom_bar(position = "dodge") +
  labs(title = "Death Count by Income and Medical Assurance",
       x = "Income",
       y = "Number of patients",
       fill = "Death status") +
  facet_wrap(~ninsclas) +
  scale_fill_manual(name = "Death", values = c( "darkolivegreen2", "red")) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
death_count

ggsave(filename = "05_death_count_1.png",
       plot = death_count,
       device = "png",
       path = "../results",
       scale = 1,
       limitsize = TRUE)

```

### How many people under $11k have private and how many have Medicare?
```{r}
income_medicalinsurance <- rhc_aug |> 
  ggplot(aes(x = income, 
             fill = factor(ninsclas))) +
  geom_bar() +
  labs(title = "Income by Medical Assurance",
       x = "Income",
       y = "Number of Patients",
       fill = "Medical Insurance")

income_medicalinsurance

ggsave(filename = "05_income_medicalinsurance_2.png",
       plot = income_medicalinsurance,
       device = "png",
       path = "../results",
       scale = 1,
       limitsize = TRUE)
```
```{r}
rhc_aug |> 
  filter(income == "Under $11k", 
         ninsclas %in% c("Medicare", "Private")) |> 
  count(income, ninsclas)
```


```{r}
# Bar plot for RHC by income and race

rhc_aug |>
  # filter(swang1 == 'RHC') |> 
  ggplot(aes(x = race,
             fill = as.factor(swang1))) +
  geom_bar(position = "dodge") +
  labs(title = "RHC Count by Income and Race",
       x = "Income",
       y = "Number of patients") +
  facet_wrap(~income) +
  theme(axis.text.x = element_text(angle = 45, 
                                   hjust = 1)) +
  theme_minimal()
  
```

```{r}
# Bar plot for RHC by income and medical assurance
rhc_aug |>
  ggplot(aes(x = race,
             fill = as.factor(swang1),
             group = as.factor(income))) +
  geom_bar(position = "dodge") +
  labs(title = "RHC Count by Income and Medical Assurance",
       x = "Income",
       y = "Count",
       fill = "RHC status") +
  facet_wrap(~income) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
```
